Revert "Implement login"
