--- EventPlanner/src/app/infrastructure/auth/auth.service.ts
@@ -1,48 +1,0 @@
-import { Injectable } from '@angular/core';
-import {HttpClient, HttpHeaders} from '@angular/common/http';
-import {BehaviorSubject, Observable} from 'rxjs';
-import {environment} from '../../../env/environment';
-import {LoginRequestDto} from './model/login-request-dto.model';
-import {JwtHelperService} from '@auth0/angular-jwt';
-import {LoginResponseDTO} from './model/login-response-dto.model';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class AuthService {
-
-  private headers = new HttpHeaders({
-    'Content-Type': 'application/json',
-    skip: 'true',
-  });
-
-  user$ = new BehaviorSubject("");
-  userState = this.user$.asObservable();
-
-  constructor(private http: HttpClient) {
-    this.user$.next(this.getRole());
-  }
-
-  login(auth: LoginRequestDto): Observable<LoginResponseDTO> {
-    return this.http.post<LoginResponseDTO>(environment.apiHost + '/auth/login', auth, {
-      headers: this.headers,
-    });
-  }
-
-  getRole(): any {
-    if (this.isLoggedIn()) {
-      const accessToken: any = localStorage.getItem('user');
-      const helper = new JwtHelperService();
-      return helper.decodeToken(accessToken).role;
-    }
-    return null;
-  }
-
-  isLoggedIn(): boolean {
-    return localStorage.getItem('user') != null;
-  }
-
-  setUser(): void {
-    this.user$.next(this.getRole());
-  }
-}
