--- EventPlanner/src/app/infrastructure/auth/interceptor.ts
@@ -1,23 +1,0 @@
-import {Injectable} from '@angular/core';
-import {HttpEvent, HttpHandler, HttpInterceptor, HttpRequest} from '@angular/common/http';
-import {Observable} from 'rxjs';
-
-@Injectable()
-export class Interceptor implements HttpInterceptor {
-  intercept(
-    req: HttpRequest<any>,
-    next: HttpHandler
-  ): Observable<HttpEvent<any>> {
-    const accessToken: any = localStorage.getItem('user');
-    if (req.headers.get('skip')) return next.handle(req);
-
-    if (accessToken) {
-      const cloned = req.clone({
-        headers: req.headers.set('Authorization', "Bearer " + accessToken),
-      });
-      return next.handle(cloned);
-    } else {
-      return next.handle(req);
-    }
-  }
-}
