--- EventPlanner/src/app/infrastructure/auth/model/login-request-dto.model.ts
@@ -1,4 +1,0 @@
-export interface LoginRequestDto {
-  email: string;
-  password: string;
-}
