--- EventPlanner/src/app/infrastructure/auth/model/login-response-dto.model.ts
@@ -1,3 +1,0 @@
-export interface LoginResponseDTO {
-  accessToken: string;
-}
