--- EventPlanner/src/app/layout/admin-nav-bar-menu/admin-nav-bar-menu.component.html
@@ -1,16 +1,0 @@
-<button mat-menu-item [routerLink]="['event-types']">
-  <mat-icon>event</mat-icon>
-  <span>Event Type Management</span>
-</button>
-<button mat-menu-item [routerLink]="['notification-panel']">
-  <mat-icon>notifications</mat-icon>
-  <span>Notification Management</span>
-</button>
-<button mat-menu-item [routerLink]="['offering-categories']">
-  <mat-icon>category</mat-icon>
-  <span>Category Management</span>
-</button>
-<button mat-menu-item>
-  <mat-icon>chat</mat-icon>
-  <span>Chat</span>
-</button>
