--- EventPlanner/src/app/layout/admin-nav-bar-menu/admin-nav-bar-menu.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-admin-nav-bar-menu',
-  templateUrl: './admin-nav-bar-menu.component.html',
-  styleUrl: './admin-nav-bar-menu.component.css'
-})
-export class AdminNavBarMenuComponent {
-
-}
