--- EventPlanner/src/app/layout/authenticated-user-nav-bar-menu/authenticated-user-nav-bar-menu.component.html
@@ -1,8 +1,0 @@
-<button mat-menu-item [routerLink]="['notification-panel']">
-  <mat-icon>notifications</mat-icon>
-  <span>Notification Management</span>
-</button>
-<button mat-menu-item>
-  <mat-icon>chat</mat-icon>
-  <span>Chat</span>
-</button>
