--- EventPlanner/src/app/layout/authenticated-user-nav-bar-menu/authenticated-user-nav-bar-menu.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-authenticated-user-nav-bar-menu',
-  templateUrl: './authenticated-user-nav-bar-menu.component.html',
-  styleUrl: './authenticated-user-nav-bar-menu.component.css'
-})
-export class AuthenticatedUserNavBarMenuComponent {
-
-}
