+++ EventPlanner/src/app/layout/login/login.component.ts
@@ -1,0 +1,22 @@
+import { Component } from '@angular/core';
+import {FormControl, FormGroup, Validators} from '@angular/forms';
+import {UserService} from '../../user/user.service';
+
+@Component({
+  selector: 'app-login',
+  templateUrl: './login.component.html',
+  styleUrl: './login.component.css'
+})
+export class LoginComponent {
+  loginForm = new FormGroup({
+    email: new FormControl('', [Validators.required, Validators.email]),
+    password: new FormControl('', [Validators.required, Validators.minLength(8)])});
+
+  constructor(private userService: UserService) {}
+
+  login(){
+    if(!this.loginForm.valid)
+      return;
+    this.userService.login(this.loginForm.value.email, this.loginForm.value.password);
+  }
+}
