--- EventPlanner/src/app/layout/organizer-nav-bar-menu/organizer-nav-bar-menu.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-organizer-nav-bar-menu',
-  templateUrl: './organizer-nav-bar-menu.component.html',
-  styleUrl: './organizer-nav-bar-menu.component.css'
-})
-export class OrganizerNavBarMenuComponent {
-
-}
