--- EventPlanner/src/app/layout/provider-nav-bar-menu/provider-nav-bar-menu.component.html
@@ -1,16 +1,0 @@
-<button mat-menu-item [routerLink]="['manage-offerings']" >
-  <mat-icon>shopping_cart</mat-icon>
-  <span>Offering Management</span>
-</button>
-<button mat-menu-item>
-  <mat-icon>money</mat-icon>
-  <span>Pricelist</span>
-</button>
-<button mat-menu-item [routerLink]="['notification-panel']">
-  <mat-icon>notifications</mat-icon>
-  <span>Notification Management</span>
-</button>
-<button mat-menu-item>
-  <mat-icon>chat</mat-icon>
-  <span>Chat</span>
-</button>
