--- EventPlanner/src/app/layout/provider-nav-bar-menu/provider-nav-bar-menu.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-provider-nav-bar-menu',
-  templateUrl: './provider-nav-bar-menu.component.html',
-  styleUrl: './provider-nav-bar-menu.component.css'
-})
-export class ProviderNavBarMenuComponent {
-
-}
